refactor Bifrost tests and remove duplicate mock
