--- src/__tests__/bifrost.test.ts
@@ -1,619 +1,0 @@
-// Skip importing the actual module to avoid ESM issues
-// We'll fully mock the module's functions
-
-/**
- * Testing strategy for Bifrost:
- * 1. We completely mock the bifrost module to avoid ESM import issues
- * 2. We define TypeScript interfaces to maintain type safety
- * 3. We implement realistic mock behavior to simulate the actual functions
- * 4. We test for both success cases and error conditions
- */
-
-// Define the types we need for testing
-interface GroupPackage {
-  threshold: number;
-  commits: CommitPackage[];
-  group_pk: string;
-}
-
-interface CommitPackage {
-  idx: number;
-  pubkey: string;
-  hidden_pn: string;
-  binder_pn: string;
-}
-
-interface SharePackage {
-  idx: number;
-  binder_sn: string;
-  hidden_sn: string;
-  seckey: string;
-}
-
-// Create more sophisticated mock implementations
-const generateKeysetWithSecret = jest.fn().mockImplementation(
-  (threshold: number, totalMembers: number, secretKey: string) => {
-    // Validate input like the original function
-    if (threshold <= 0 || totalMembers <= 0) {
-      throw new Error('Threshold and total members must be positive numbers');
-    }
-    if (threshold > totalMembers) {
-      throw new Error('Threshold cannot be greater than total members');
-    }
-    if (!secretKey || typeof secretKey !== 'string') {
-      throw new Error('Secret key must be a non-empty string');
-    }
-    
-    // Add constraint for unreasonably large values
-    if (threshold > 100 || totalMembers > 100) {
-      throw new Error('Threshold and member count must be reasonable');
-    }
-    
-    // For nsec format, simulate decoding
-    let processedKey = secretKey;
-    if (secretKey.startsWith('nsec')) {
-      // Use a hash of the input key to simulate different outputs for different keys
-      processedKey = secretKey.split('').reduce((acc, char) => acc + char.charCodeAt(0), 0).toString();
-    }
-    
-    // Return a realistic result structure with outputs dependent on the inputs
-    const shareCredentials = Array(totalMembers).fill(0).map((_, i) => 
-      `mocked_share_pkg_${processedKey.substring(0, 8)}_${i+1}`
-    );
-    
-    return {
-      groupCredential: `mocked_group_pkg_${processedKey.substring(0, 8)}_${threshold}_of_${totalMembers}`,
-      shareCredentials
-    };
-  }
-);
-
-const decode_share = jest.fn().mockImplementation((share: string) => {
-  // Check for valid input
-  if (!share || typeof share !== 'string') {
-    throw new Error('Invalid share format');
-  }
-  
-  if (!share.startsWith('bfshare')) {
-    throw new Error('Invalid share format (should start with bfshare)');
-  }
-  
-  // Test for invalid format
-  if (!share.includes('1') && share !== 'bfshare1_1') {
-    throw new Error('Invalid share format');
-  }
-  
-  // Return different values based on input to simulate decoding different shares
-  const shareNumber = parseInt(share.split('_').pop() || '1');
-  
-  return {
-    idx: shareNumber,
-    binder_sn: `binder_sn_${shareNumber}`,
-    hidden_sn: `hidden_sn_${shareNumber}`,
-    seckey: `share_seckey_${shareNumber}`
-  };
-});
-
-const decode_group = jest.fn().mockImplementation((group: string) => {
-  // Check for valid input
-  if (!group || typeof group !== 'string') {
-    throw new Error('Invalid group format');
-  }
-  
-  if (!group.startsWith('bfgroup')) {
-    throw new Error('Invalid group format (should start with bfgroup)');
-  }
-  
-  // Test for invalid threshold/member count format
-  if (group === 'bfgroup1_0_of_0') {
-    throw new Error('Invalid group format');
-  }
-  
-  // Parse the threshold and total from the mock format (e.g., "mocked_group_pkg_2_of_3")
-  const parts = group.split('_');
-  
-  // Look for threshold in the format pattern
-  let threshold = 2; // Default value
-  let totalMembers = 3; // Default value
-  
-  if (group.includes('3_of_5')) {
-    threshold = 3;
-    totalMembers = 5;
-  } else if (group.includes('2_of_3')) {
-    threshold = 2;
-    totalMembers = 3;
-  } else if (parts.length >= 4 && !isNaN(parseInt(parts[3]))) {
-    // Try to extract from pattern like 'bfgroup_3_of_5'
-    threshold = parseInt(parts[3]);
-    if (parts.length >= 6 && !isNaN(parseInt(parts[5]))) {
-      totalMembers = parseInt(parts[5]);
-    }
-  }
-  
-  // Generate commits based on the total
-  const commits = Array(totalMembers).fill(0).map((_, i) => ({
-    idx: i + 1,
-    pubkey: `pubkey${i + 1}`,
-    hidden_pn: `hidden${i + 1}`,
-    binder_pn: `binder${i + 1}`
-  }));
-  
-  return {
-    threshold,
-    commits,
-    group_pk: 'group_pubkey'
-  };
-});
-
-const recover_nsec = jest.fn().mockImplementation(
-  (group: GroupPackage, shares: SharePackage[]) => {
-    // Validate inputs
-    if (!group || !shares || shares.length === 0) {
-      throw new Error('Group package and at least one share package are required');
-    }
-    
-    if (shares.length < group.threshold) {
-      throw new Error(`Not enough shares provided. Need at least ${group.threshold} shares`);
-    }
-    
-    // Validate share indices
-    for (const share of shares) {
-      if (share.idx < 0) {
-        throw new Error('Invalid share index');
-      }
-      
-      // Check for incomplete shares
-      if (!share.seckey || !share.hidden_sn || !share.binder_sn) {
-        throw new Error('Invalid share format');
-      }
-    }
-    
-    // Sort shares by index to simulate real behavior
-    const sortedShares = [...shares].sort((a, b) => a.idx - b.idx);
-    
-    // Pretend to combine shares
-    let combinedSecretParts = '';
-    for (let i = 0; i < group.threshold; i++) {
-      if (i < sortedShares.length) {
-        combinedSecretParts += sortedShares[i].idx;
-      }
-    }
-    
-    // Return a nsec that encodes which shares were used
-    return `nsec1_using_shares_${combinedSecretParts}`;
-  }
-);
-
-// Mock the entire module
-jest.mock('../lib/bifrost', () => ({
-  generateKeysetWithSecret,
-  decode_share,
-  decode_group,
-  recover_nsec
-}));
-
-// Reset all mocks before each test
-beforeEach(() => {
-  generateKeysetWithSecret.mockClear();
-  decode_share.mockClear();
-  decode_group.mockClear();
-  recover_nsec.mockClear();
-});
-
-describe('Bifrost Module Functions', () => {
-  describe('generateKeysetWithSecret', () => {
-    it('should generate a keyset with valid parameters', () => {
-      const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-      const threshold = 2;
-      const totalMembers = 3;
-      
-      const result = generateKeysetWithSecret(threshold, totalMembers, secretKey);
-      
-      // Verify function was called with correct parameters
-      expect(generateKeysetWithSecret).toHaveBeenCalledWith(threshold, totalMembers, secretKey);
-      
-      // Verify the structure of the result
-      expect(result).toHaveProperty('groupCredential');
-      expect(result).toHaveProperty('shareCredentials');
-      expect(result.shareCredentials).toHaveLength(3);
-      
-      // Verify values reflect the parameters
-      expect(result.groupCredential).toContain('2_of_3');
-      expect(result.shareCredentials[0]).toContain('1');
-      expect(result.shareCredentials[1]).toContain('2');
-      expect(result.shareCredentials[2]).toContain('3');
-    });
-    
-    it('should handle nsec input format', () => {
-      const nsecKey = 'nsec1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx';
-      const threshold = 2;
-      const totalMembers = 3;
-      
-      const result = generateKeysetWithSecret(threshold, totalMembers, nsecKey);
-      
-      // Verify function was called with correct parameters
-      expect(generateKeysetWithSecret).toHaveBeenCalledWith(threshold, totalMembers, nsecKey);
-      
-      // Verify the structure is the same as with hex key
-      expect(result).toHaveProperty('groupCredential');
-      expect(result).toHaveProperty('shareCredentials');
-      expect(result.shareCredentials).toHaveLength(3);
-    });
-    
-    it('should support different threshold configurations', () => {
-      // Test a 3-of-5 configuration
-      const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-      const threshold = 3;
-      const totalMembers = 5;
-      
-      const result = generateKeysetWithSecret(threshold, totalMembers, secretKey);
-      
-      expect(result.groupCredential).toContain('3_of_5');
-      expect(result.shareCredentials).toHaveLength(5);
-      
-      // Test a 1-of-1 configuration
-      const result2 = generateKeysetWithSecret(1, 1, secretKey);
-      
-      expect(result2.groupCredential).toContain('1_of_1');
-      expect(result2.shareCredentials).toHaveLength(1);
-    });
-    
-    it('should throw error when threshold is greater than total members', () => {
-      const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-      const threshold = 4;
-      const totalMembers = 3;
-      
-      expect(() => {
-        generateKeysetWithSecret(threshold, totalMembers, secretKey);
-      }).toThrow('Threshold cannot be greater than total members');
-    });
-    
-    it('should throw error when threshold or total members are invalid', () => {
-      const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-      
-      // Test negative threshold
-      expect(() => {
-        generateKeysetWithSecret(-1, 3, secretKey);
-      }).toThrow('Threshold and total members must be positive numbers');
-      
-      // Test zero threshold
-      expect(() => {
-        generateKeysetWithSecret(0, 3, secretKey);
-      }).toThrow('Threshold and total members must be positive numbers');
-      
-      // Test negative total members
-      expect(() => {
-        generateKeysetWithSecret(2, -1, secretKey);
-      }).toThrow('Threshold and total members must be positive numbers');
-    });
-    
-    it('should throw error when secret key is empty', () => {
-      const secretKey = '';
-      const threshold = 2;
-      const totalMembers = 3;
-      
-      expect(() => {
-        generateKeysetWithSecret(threshold, totalMembers, secretKey);
-      }).toThrow('Secret key must be a non-empty string');
-    });
-    
-    it('should throw error for very large threshold and member values', () => {
-      const secretKey = 'd5511c9ab0ef3578c939da718bf7bcfebd3a8bacce7cd7e58dcd7d8bf9aa374e';
-      const threshold = 1000; // Unreasonably large threshold
-      const totalMembers = 2000; // Unreasonably large member count
-      
-      expect(() => {
-        generateKeysetWithSecret(threshold, totalMembers, secretKey);
-      }).toThrow('Threshold and member count must be reasonable');
-    });
-
-    it('should handle various nsec formats', () => {
-      // Test with different nsec formats
-      const validNsec1 = 'nsec1xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx';
-      const validNsec2 = 'nsec1yyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyyy';
-      
-      // Both should work and produce different results
-      const result1 = generateKeysetWithSecret(2, 3, validNsec1);
-      const result2 = generateKeysetWithSecret(2, 3, validNsec2);
-      
-      expect(result1.groupCredential).not.toBe(result2.groupCredential);
-    });
-  });
-  
-  describe('decode_share', () => {
-    it('should decode a share package', () => {
-      const share = 'bfshare1_1';
-      
-      const result = decode_share(share);
-      
-      // Verify function was called correctly
-      expect(decode_share).toHaveBeenCalledWith(share);
-      
-      // Verify the structure of the result
-      expect(result).toEqual({
-        idx: 1,
-        binder_sn: 'binder_sn_1',
-        hidden_sn: 'hidden_sn_1',
-        seckey: 'share_seckey_1'
-      });
-    });
-    
-    it('should decode different shares with different values', () => {
-      const share1 = 'bfshare1_1';
-      const share2 = 'bfshare1_2';
-      
-      const result1 = decode_share(share1);
-      const result2 = decode_share(share2);
-      
-      // Results should be different based on share identity
-      expect(result1.idx).toBe(1);
-      expect(result2.idx).toBe(2);
-      expect(result1.seckey).not.toBe(result2.seckey);
-    });
-    
-    it('should throw error for invalid share format', () => {
-      // Test empty share
-      expect(() => {
-        decode_share('');
-      }).toThrow('Invalid share format');
-      
-      // Test invalid prefix
-      expect(() => {
-        decode_share('invalid_prefix');
-      }).toThrow('Invalid share format');
-    });
-    
-    it('should throw proper errors for malformed shares', () => {
-      // Test with undefined input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_share(undefined);
-      }).toThrow('Invalid share format');
-      
-      // Test with null input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_share(null);
-      }).toThrow('Invalid share format');
-      
-      // Test with non-string input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_share(123);
-      }).toThrow('Invalid share format');
-    });
-    
-    it('should reject shares with invalid format but correct prefix', () => {
-      expect(() => {
-        decode_share('bfshare_but_invalid_format');
-      }).toThrow('Invalid share format');
-    });
-  });
-  
-  describe('decode_group', () => {
-    it('should decode a group into a structured object', () => {
-      const group = 'bfgroup_mock';
-      
-      const result = decode_group(group);
-      
-      // Verify function was called with correct parameters
-      expect(decode_group).toHaveBeenCalledWith(group);
-      
-      // Verify the structure of the result
-      expect(result).toHaveProperty('threshold');
-      expect(result).toHaveProperty('commits');
-      expect(result).toHaveProperty('group_pk');
-    });
-
-    it('should validate group package structure with correct threshold', () => {
-      const mockGroup = 'bfgroup_3_of_5';
-      
-      const result = decode_group(mockGroup);
-      
-      expect(result).toEqual({
-        threshold: 3,
-        commits: expect.arrayContaining([
-          expect.objectContaining({
-            idx: expect.any(Number),
-            pubkey: expect.any(String),
-            hidden_pn: expect.any(String),
-            binder_pn: expect.any(String)
-          })
-        ]),
-        group_pk: expect.any(String)
-      });
-      
-      // Verify threshold is correctly set to 3
-      expect(result.threshold).toBe(3);
-      expect(result.commits.length).toBe(5);
-    });
-    
-    it('should decode different groups with appropriate member counts', () => {
-      const group1 = 'bfgroup1_2_of_3';
-      const group2 = 'bfgroup1_3_of_5';
-      
-      const result1 = decode_group(group1);
-      const result2 = decode_group(group2);
-      
-      expect(result1.threshold).toBe(2);
-      expect(result1.commits.length).toBe(3);
-      
-      expect(result2.threshold).toBe(3);
-      expect(result2.commits.length).toBe(5);
-    });
-    
-    it('should throw error for invalid group format', () => {
-      // Test empty group
-      expect(() => {
-        decode_group('');
-      }).toThrow('Invalid group format');
-      
-      // Test invalid prefix
-      expect(() => {
-        decode_group('invalid_prefix');
-      }).toThrow('Invalid group format');
-    });
-    
-    it('should handle invalid input types gracefully', () => {
-      // Test with undefined input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_group(undefined);
-      }).toThrow('Invalid group format');
-      
-      // Test with null input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_group(null);
-      }).toThrow('Invalid group format');
-      
-      // Test with object input
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        decode_group({});
-      }).toThrow('Invalid group format');
-    });
-    
-    it('should handle groups with invalid member counts', () => {
-      // Mock implementation should check thresholds and handle this case
-      expect(() => {
-        decode_group('bfgroup1_0_of_0'); // Invalid threshold and member count
-      }).toThrow('Invalid group format');
-    });
-  });
-  
-  describe('recover_nsec', () => {
-    it('should recover a secret key from group and shares', () => {
-      const group = {
-        threshold: 2,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      const shares = [
-        { idx: 1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' },
-        { idx: 2, binder_sn: 'binder2', hidden_sn: 'hidden2', seckey: 'seckey2' }
-      ];
-      
-      const result = recover_nsec(group, shares);
-      
-      // Verify function was called correctly
-      expect(recover_nsec).toHaveBeenCalledWith(group, shares);
-      
-      // Check the result value reflects the shares used
-      expect(result).toContain('using_shares');
-      expect(result).toContain('12'); // Share indices 1,2
-    });
-    
-    it('should work with more shares than needed', () => {
-      const group = {
-        threshold: 2,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      const shares = [
-        { idx: 1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' },
-        { idx: 2, binder_sn: 'binder2', hidden_sn: 'hidden2', seckey: 'seckey2' },
-        { idx: 3, binder_sn: 'binder3', hidden_sn: 'hidden3', seckey: 'seckey3' }
-      ];
-      
-      const result = recover_nsec(group, shares);
-      
-      // Still works with extra shares provided
-      expect(result).toContain('using_shares');
-      expect(result).toContain('12'); // First 2 shares used (threshold is 2)
-    });
-    
-    it('should handle shares provided in any order', () => {
-      const group = {
-        threshold: 2,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      // Provide shares in reverse order
-      const shares = [
-        { idx: 3, binder_sn: 'binder3', hidden_sn: 'hidden3', seckey: 'seckey3' },
-        { idx: 2, binder_sn: 'binder2', hidden_sn: 'hidden2', seckey: 'seckey2' },
-        { idx: 1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' }
-      ];
-      
-      const result = recover_nsec(group, shares);
-      
-      // Should always use the first 'threshold' number of shares after sorting
-      expect(result).toContain('using_shares');
-      expect(result).toContain('12'); // Lowest 2 indices after sorting
-    });
-    
-    it('should throw error when not enough shares are provided', () => {
-      const group = {
-        threshold: 3,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      const shares = [
-        { idx: 1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' },
-        { idx: 2, binder_sn: 'binder2', hidden_sn: 'hidden2', seckey: 'seckey2' }
-      ];
-      
-      expect(() => {
-        recover_nsec(group, shares);
-      }).toThrow('Not enough shares provided');
-    });
-    
-    it('should throw error when group is missing', () => {
-      expect(() => {
-        // @ts-ignore - Intentionally passing incorrect params for test
-        recover_nsec(null, [
-          { idx: 1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' }
-        ]);
-      }).toThrow('Group package and at least one share package are required');
-    });
-    
-    it('should throw error when shares array is empty', () => {
-      expect(() => {
-        recover_nsec({threshold: 2, commits: [], group_pk: 'pk'}, []);
-      }).toThrow('Group package and at least one share package are required');
-    });
-    
-    it('should validate share index values', () => {
-      const group = {
-        threshold: 2,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      // Test with invalid share indices (negative)
-      const sharesWithNegativeIdx = [
-        { idx: -1, binder_sn: 'binder1', hidden_sn: 'hidden1', seckey: 'seckey1' },
-        { idx: -2, binder_sn: 'binder2', hidden_sn: 'hidden2', seckey: 'seckey2' }
-      ];
-      
-      expect(() => {
-        recover_nsec(group, sharesWithNegativeIdx);
-      }).toThrow('Invalid share index');
-    });
-    
-    it('should reject invalid share content', () => {
-      const group = {
-        threshold: 2,
-        commits: [],
-        group_pk: 'group_pk'
-      };
-      
-      // Test with missing required properties
-      const incompleteShares = [
-        // @ts-ignore - Intentionally passing incorrect params for test
-        { idx: 1 }, // Missing other required fields
-        // @ts-ignore - Intentionally passing incorrect params for test
-        { idx: 2, binder_sn: 'binder2' } // Missing hidden_sn and seckey
-      ];
-      
-      expect(() => {
-        recover_nsec(group, incompleteShares);
-      }).toThrow('Invalid share format');
-    });
-  });
-});
