--- src/__tests__/mocks/.gitkeep
@@ -1,2 +1,0 @@
-# This file exists to make sure the mocks directory is included in Git
-# It doesn't contain any actual code
